fix: getDayOfWeek fuction move to util
fix: getDayOfWeek fuction move to util
