+++ feature/club/src/main/java/in/koreatech/koin/feature/club/ui/clubrecruitcreate/ClubRecruitCreateScreen.kt
@@ -60,7 +60,7 @@ import `in`.koreatech.koin.feature.club.component.KoinClubDatePickerDialog
 import `in`.koreatech.koin.feature.club.component.KoinClubDateSelectBox
 import `in`.koreatech.koin.feature.club.component.KoinClubExtraSmallDialog
 import `in`.koreatech.koin.feature.club.component.KoinClubExtraSmallDialogDanger
-import `in`.koreatech.koin.feature.club.component.getDayOfWeek
+import `in`.koreatech.koin.feature.club.utils.getDayOfWeek
 import `in`.koreatech.koin.feature.club.utils.pickMedia
 import java.time.LocalDate
 import org.orbitmvi.orbit.compose.collectAsState
