+++ feature/club/src/main/java/in/koreatech/koin/feature/club/utils/DayOfWeekUtil.kt
@@ -1,0 +1,14 @@
+package `in`.koreatech.koin.feature.club.utils
+
+import java.time.DayOfWeek
+
+fun getDayOfWeek(dayOfWeek: DayOfWeek): String =
+    when (dayOfWeek) {
+        DayOfWeek.MONDAY -> "월"
+        DayOfWeek.TUESDAY -> "화"
+        DayOfWeek.WEDNESDAY -> "수"
+        DayOfWeek.THURSDAY -> "목"
+        DayOfWeek.FRIDAY -> "금"
+        DayOfWeek.SATURDAY -> "토"
+        DayOfWeek.SUNDAY -> "일"
+    }
